pyproject.toml
[project]
name = "midas"
version = "0.1.0"
description = "AI trading assistant — Phase 1 local demo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.100.0",
    "argon2-cffi>=25.1.0",
    "arq>=0.28.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "curl-cffi>=0.15.0",
    "email-validator>=2.3.0",
    "fastapi>=0.136.1",
    "feedparser>=6.0.12",
    "httpx>=0.28.1",
    "lxml>=6.1.0",
    "nselib>=2.5.1",
    "openai>=2.35.1",
    "pandas>=3.0.2",
    "pandas-ta-classic>=0.5.44",
    "pydantic>=2",
    "pydantic-settings>=2.14.0",
    "pyjwt[crypto]>=2.12.1",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "redis[hiredis]>=5.3.1",
    "sqlalchemy[asyncio]>=2.0",
    "sse-starlette>=3.4.2",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.46.0",
    "vadersentiment>=3.3.2",
    "yfinance>=1.3.0",
]

[dependency-groups]
dev = [
    "mypy>=2.0.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.12",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_mode = "auto"
